The start of putting shit into modules
Made a struct called Chunk in its own file. This struct will contain all the Chunk functionality of course. Also created a new file for the Vertex data, so I can set it up and still pass it to the chunk.rs file, as it's not created in a normal fashion.